Implement Taker State Machine
