+++ .env
@@ -1,11 +1,11 @@
-BITCOIN_HD_KEY_0=tprv8ZgxMBicQKsPdgMSaE4EdiPPHiyh8gZuKMvTPzko6avGsNV2AtajH6yCYZp4ANrgKjT4jbPTUAgK3tN5314fgtgJtJGB9HKmSDsRFkVnZ9b
-BITCOIN_HD_KEY_1=tprv8ZgxMBicQKsPeSZbbGaxnij9SDeufRqNTcJQBedPxTnRmHabzzxo6HSkd68ymSNaMyJZu2ZSUpuXRvJULa7JZDVnWMZUiS2daBRmfcX7qEb
-BITCOIN_P2P_URI=127.0.0.1:54013
+BITCOIN_HD_KEY_0=tprv8ZgxMBicQKsPeAEQCpu2pgEXJvMVroWmt4oJU8dEmBXmPuEqpi71QHeeuzEcjRUgUE9MiWRhNzcYhdwsmdTSVeERSBJY5CVxobyNvfo2tz1
+BITCOIN_HD_KEY_1=tprv8ZgxMBicQKsPeA58wvStUN6KnA3VGjK9BAbbBT6oVSFAENXSUc8dDgdLf4ecGo5um7XiP3oJ5kY9MbEoe8wkH5YK7gfwVPdsQRPQ8QGNor7
+BITCOIN_P2P_URI=127.0.0.1:62938
 ERC20_CONTRACT_ADDRESS=0x62d69f6867a0a084c6d313943dc22023bc263691
-ETHEREUM_KEY_0=e887c235d974e8315d107ab1cfb49bfba37e11396b2dac5a1eed43bb3913a3ca
-ETHEREUM_KEY_1=67ae89e07cb8e5f65b615d42d89b9d77715d79ed7a0fa1780fa695b4e0775b63
-ETHEREUM_NODE_HTTP_URL=http://127.0.0.1:53997
-HTTP_URL_CND_0=http://127.0.0.1:54026
-HTTP_URL_CND_1=http://127.0.0.1:54030
+ETHEREUM_KEY_0=e0079a62bd9176f541461601e3a7449e7dadc5d133d543f3308a92ee2c52e037
+ETHEREUM_KEY_1=be13bd25f1e037b556b426d8b735ee587dbc0cb27695b4aaece18936d63ceca3
+ETHEREUM_NODE_HTTP_URL=http://127.0.0.1:62926
+HTTP_URL_CND_0=http://127.0.0.1:62947
+HTTP_URL_CND_1=http://127.0.0.1:62951
 MAKER_URL=http://localhost:3000
 POLL_INTERVAL=5000
